ok i can commit now thank god
