+++ src/python/grapl-common/grapl_common/env_helpers.py
@@ -225,3 +225,7 @@ class SecretsManagerClientFactory(FromEnv["SecretsManagerClient"]):
             self.client_create_fn, _SecretsManagerParams
         )
         return client
+
+
+def get_deployment_name() -> str:
+    return os.environ["DEPLOYMENT_NAME"]
