+++ src/python/grapl_e2e_tests/BUILD
@@ -2,5 +2,5 @@ python_library()
 
 python_tests(
   name="tests",
-  sources=["tests.py"]
+  sources=["**/*.py"]
 )
