--- src/python/grapl_e2e_tests/expected_gql_asset.py
@@ -1,43 +1,0 @@
-from typing import Any, Mapping
-
-
-def expected_gql_asset() -> Mapping[str, Any]:
-    """
-    All the fixed values (i.e. no uid, no node key) we'd see in the e2e test
-    """
-    return {
-        "dgraph_type": ["Asset"],
-        "display": "DESKTOP-FVSHABR",
-        "hostname": "DESKTOP-FVSHABR",
-        "asset_processes": [
-            {
-                "dgraph_type": ["Process"],
-                "process_name": "cmd.exe",
-                "process_id": 5824,
-            },
-            {
-                "dgraph_type": ["Process"],
-                "process_name": "dropper.exe",
-                "process_id": 4164,
-            },
-            {
-                "dgraph_type": ["Process"],
-                "process_name": "cmd.exe",
-                "process_id": 5824,
-            },
-            {
-                "dgraph_type": ["Process"],
-                "process_name": "svchost.exe",
-                "process_id": 6132,
-            },
-        ],
-        "files_on_asset": None,
-        "risks": [
-            {
-                "dgraph_type": ["Risk"],
-                "node_key": "Rare Parent of cmd.exe",
-                "analyzer_name": "Rare Parent of cmd.exe",
-                "risk_score": 10,
-            }
-        ],
-    }
